+++ src/utils.js
@@ -17,7 +17,9 @@ if (window.matchMedia) {
 }
 
 
-
+export function deepCopy (o) {
+	return structuredClone(o);
+}
 
 
 // native js animation
@@ -132,4 +134,27 @@ export function blink (el, duration = 160) {
 		{ opacity: 0.5 },
 		{ duration: duration / 2, fill: 'backwards' }
 	);
+}
+
+
+export function timeAgo (date, now) {
+	now = now || new Date().getTime();
+	let seconds = (now - +date) / 1000;
+	const intervals = [
+		{ label: 'year', seconds: 31536000 },
+		{ label: 'month', seconds: 2592000 },
+		{ label: 'day', seconds: 86400 },
+		{ label: 'hour', seconds: 3600 },
+		{ label: 'minute', seconds: 60 }
+	];
+	const chunks = [];
+	while (seconds > 60) {
+		const interval = intervals.find(i => i.seconds < seconds) || { seconds: 0, label: '' };
+		const count = Math.floor(seconds / (interval.seconds || seconds));
+		chunks.push(`${count} ${interval.label}${count !== 1 ? 's' : ''}`);
+		seconds -= count * (interval.seconds || seconds);
+	}
+	if (!chunks.length) return 'just now';
+	if (chunks.length === 1) return chunks[0] + ' ago';
+	return;
 }